Move Files to Auth Ownership
