+++ apps/web/src/app/settings/security-routing.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from "@angular/core";
 import { RouterModule, Routes } from "@angular/router";
 
+import { ChangePasswordComponent } from "../auth/settings/change-password.component";
 import { TwoFactorSetupComponent } from "../auth/settings/two-factor-setup.component";
 
-import { ChangePasswordComponent } from "./change-password.component";
 import { SecurityKeysComponent } from "./security-keys.component";
 import { SecurityComponent } from "./security.component";
