+++ libs/common/src/auth/models/response/protected-device.response.ts
